Change Lat and Long to doubles
BigDecimal was not working out. Also, there was literally no need since both of these values have real world caps so we can never expect them to be much bigger than a couple digits

+++ src/main/java/com/dladeji/earthquake/dtos/QuakeDto.java
@@ -1,12 +1,10 @@
 package com.dladeji.earthquake.dtos;
 
-import java.math.BigDecimal;
-
 public interface QuakeDto {
     String getTitle();
     double getMag();
     String getPlace();
-    BigDecimal getLatitude();
-    BigDecimal getLongtitude();
+    double getLatitude();
+    double getLongtitude();
     // depth needed??
 }
