+++ src/main/resources/db/migration/V2__add_location_fields_to_table.sql
@@ -1,4 +1,4 @@
 alter table earthquakes
-    ADD COLUMN `longtitude` DECIMAL(10,4) NOT NULL,
-    ADD COLUMN `latitude` DECIMAL(10,4) NOT NULL,
+    ADD COLUMN `longtitude` DOUBLE NOT NULL,
+    ADD COLUMN `latitude` DOUBLE NOT NULL,
     ADD COLUMN `depth` DOUBLE NOT NULL;
